gdb-testsuite-don-t-error-when-trying-to-unset-last_spawn_tty_name
In spawn_capture_tty_name (lib/gdb.exp) we either set or unset
last_spawn_tty_name depending on whether spawn_out(slave,name) exists
or not.

One situation that might cause spawn_out(slave,name) to not exists is
if the spawn function is called with the argument -leaveopen, which is
how it is called when processes are created as part of a pipeline, the
created process has no tty, instead its output is written to a file
descriptor.

If a pipeline is created consisting of multiple processes then there
will be multiple sequential calls to spawn, all using -leaveopen.  The
first of these calls is fine, spawn_out(slave,name) is not set, and so
in spawn_capture_tty_name we unset last_spawn_tty_name.  However, on
the second call to spawn, spawn_out(slave,name) is still not set and
so in spawn_capture_tty_name we again try to unset
last_spawn_tty_name, this now throws an error (as last_spawn_tty_name
is already unset).

Fix this issue by using -nocomplain with the call to unset in
spawn_capture_tty_name.

Before this commit I was seeing gdb.base/gnu-debugdata.exp report 1
pass, and 1 unsupported test.  After this commit I now see 16 passes
from this test script.

I have also improved the code that used to do this:

    if { [info exists spawn_out] } {
	set ::last_spawn_tty_name $spawn_out(slave,name)
    } else {
        ...
    }

The problem here is that we check for the existence of spawn_out, and
then unconditionally read spawn_out(slave,name).  A situation could
arise where some other element of spawn_out is set,
e.g. spawn_out(foo), in which case we would enter the if block and try
to read a non-existent variable.  After this commit we now check
specifically for spawn_out(slave,name).

Finally, it is worth noting that before this issue was fixed runtest
itself, or rather the expect process behind runtest, would segfault
while exiting.  I haven't looked at all into what the problem is here
that caused expect to crash, as fixing the bug in GDB's testing
scripts made the segfault go away.