[gdb/symtab] Work around fsanitize=address false positive for per_cu->lang
When building gdb with -fsanitize=thread and gcc 12, and running test-case
gdb.dwarf2/dwz.exp, we run into a data race between:
...
  Read of size 1 at 0x7b200000300d by thread T2:^M
    #0 cutu_reader::cutu_reader(dwarf2_per_cu_data*, dwarf2_per_objfile*, \
    abbrev_table*, dwarf2_cu*, bool, abbrev_cache*) gdb/dwarf2/read.c:6164 \
    (gdb+0x82ec95)^M
...
and:
...
  Previous write of size 1 at 0x7b200000300d by main thread:^M
    #0 prepare_one_comp_unit gdb/dwarf2/read.c:23588 (gdb+0x86f973)^M
...

In other words, between:
...
  if (this_cu->reading_dwo_directly)
...
and:
...
    cu->per_cu->lang = pretend_language;
...

Both fields are part of the same bitfield, and writing to one field while
reading from another is not a problem, so this is a false positive.

An easy way to get rid of the false positive when compiling with thread
sanitizer is to do this:
...
  #ifdef __SANITIZE_THREAD__
    language lang;
  #else
    ENUM_BITFIELD (language) lang : LANGUAGE_BITS;
  #endif
...
but that also inhibits the detection of parallel writing to different fields
in the same bitfield, which is a problem.

Fix this instead by moving the lang field out of the bitfield.

In the bitfield, storing the lang field required LANGUAGE_BITS == 5 bits.

Set the underlying type of enum lang to char, to require only 8 bits outside
the bitfield.

Due to a compilation error with gcc 7.5.0, that also requires us to set
LANGUAGE_BITS to 8.

The size of struct dwarf2_per_cu_data remains the same (at least for -m64).

Tested on x86_64-linux.