[gdb/symtab] Work around fsanitize=address false positive for per_cu->unit_type
When building gdb with -fsanitize=thread and gcc 12, and running test-case
gdb.dwarf2/dwz.exp, we run into a data race between:
...
  Write of size 1 at 0x7b200000300e by thread T4:
    #0 process_psymtab_comp_unit gdb/dwarf2/read.c:6789 (gdb+0x830720)
...
and:
...
  Previous read of size 1 at 0x7b200000300e by main thread:
    #0 cutu_reader::cutu_reader(dwarf2_per_cu_data*, dwarf2_per_objfile*, \
    abbrev_table*, dwarf2_cu*, bool, abbrev_cache*) gdb/dwarf2/read.c:6164 \
    (gdb+0x82edab)
...

In other words, between:
...
      this_cu->unit_type = DW_UT_partial;
...
and:
...
  if (this_cu->reading_dwo_directly)
...

Both fields are part of the same bitfield, and writing to one field while
reading from another is not a problem, so this is a false positive.

Fix this by moving the unit_type field out of the bitfield.

Use type unsigned char instead of enum dwarf_unit_type to keep requiring only
8 bits.

The size of struct dwarf2_per_cu_data remains the same (at least for -m64).

Tested on x86_64-linux.