gdb-testsuite-ada-pie
[gdb/testsuite] Compile ada hello world for skip_ada_tests

For openSUSE Leap 15.0 with gcc-PIE installed (which makes gcc create PIE
executables by default) we get:
...
FAIL: gdb.ada/O2_float_param.exp: compilation foo.adb
...

The problem is that while gcc-PIE affects gcc, it does not affect gnatlink,
so it links in the libgnat.a, rather than libgnat_pic.a.  [ This is
bsc#1115034. ]

[ Without gcc-PIE, we have a related problem: if we run ada tests with
--target_board=unix/-fPIE/-pie, which makes sure PIE executables are generated
for c/c++ test-cases, still we get non-PIE ada executables, because gnatmake
does not pass -pie to gnatlink.  And if gnatmake would pass -pie to gnatlink,
we'd run into the same FAIL as above because gnatlink does not use use
libgnat_pic.a when -pie is specified (this is PR gcc/87936).  So, in order to
have ada tests generate PIE executables, we need
--target_board=unix/-fPIE/-largs/-pie/-lgnat_pic/-margs, which will not work
with c/c++ test-cases.  ]

For now, we check whether we can compile an ada hello world, and if not,
generate an UNSUPPORTED in either skip_ada_tests or gdb_compile_ada, to
not have this problem result in ~200 FAILs.

gdb/testsuite/ChangeLog:

2020-11-06  Tom de Vries  <tdevries@suse.de>

	* lib/ada.exp (gdb_compile_ada): Call gdb_can_compile_ada.
	(gdb_can_compile_ada): New gdb_caching_proc.
	* lib/gdb.exp: Add load_lib ada.exp.
	(skip_ada_tests): Return 1 if !gdb_can_compile_ada.