[gdb/build] Add gdb/contrib/makeinfo-dummy.sh
Currently, we cannot build gdb without makeinfo installed.

It would be convenient to work around this by using the configure flag
MAKEINFO=/usr/bin/true or some such, but that doesn't work because top-level
configure requires a makeinfo of at least version 4.7, and that version check
fails for /usr/bin/true, so we end up with MAKEINFO=missing instead.

Work around this by adding a script gdb/contrib/makeinfo-dummy.sh that can be
used instead, like so:
...
$ ./src/configure MAKEINFO=$src/gdb/contrib/makeinfo-dummy.sh
...

The script merely prints the version string that satisfies the version check
in $src/configure.

Tested on x86_64-linux, with makeinfo removed.