[gdb/symtab] Redo "Fix assert in set_length"
This reverts commit 1c04f72368c ("[gdb/symtab] Fix assert in set_length"), due
to a regression reported in PR29572, and implements a different fix for PR29453.

The fix is to not use the CU table in a .debug_names section to construct
all_comp_units, but instead use create_all_comp_units, and then verify the CU
table from .debug_names.  This also fixes PR25969, so remove the KFAIL.

Bug: https://sourceware.org/bugzilla/show_bug.cgi?id=29572
Bug: https://sourceware.org/bugzilla/show_bug.cgi?id=25969