[RFC][gdb/symtab] Add maint set symbol-read-order
With the test-case included in this patch, we run into:
...
(gdb) print var_dwarf^M
$1 = 1^M
(gdb) PASS: gdb.ctf/dwarf2-and-ctf.exp: print var_dwarf
print var_ctf^M
'var_ctf' has unknown type; cast it to its declared type^M
(gdb) FAIL: gdb.ctf/dwarf2-and-ctf.exp: print var_ctf
...

The problem is that the executable contains both ctf and dwarf2, so the ctf
info is ignored.

GDB has support for handling multiple debug formats, but the common use case
for ctf is to be used when dwarf2 is not present, and gdb reflects that,
assuming that by reading ctf in addition there won't be any extra information,
so it's not worth the additional cycles and memory.

Add a new maintenance command "maint set symbol-read-order <string>", where
the argument string "mdebug+stabs+dwarf2/ctf" represent the current default,
in other words:
- first mdebug is read
- then stabs is read
- then dwarf2 is read,
- then ctf is read, but only if dwarf2 info was missing.

A bit of background on the relevance of reading order: the formats that create
partial symtabs (all but dwarf2) have a priority relationship between them,
where reading earlier means lower priority.  By reading the format with the
most detail last, we ensure it has the highest priority, which makes sure that
in case there is overlapping info, the most detailed info is found.  This
explains the ordering of mdebug, stabs and dwarf2 (because it used to have
partial symtabs).

By using "maint set symbol-read-order ctf+dwarf2 or dwarf2+ctf, we can make the
test-case pass.

This is an RFC at this point, so no documentation yet.

Points I'd like to have feedback on:
- I chose this to be a maintenance command, because I suspect the reading
  order implications are a bit hard to understand.
- was it a good idea to use strtok ?
- I went for genericity, but could haved used instead a simple
  "set read-ctf-symtab <no|no-dwarf2|yes>".  If that is better, should
  it be a maintenance command or not?
- I initially went for the priority first order in the argument string, so
  reflecting the current default would be "dwarf2/ctf+stabs+mdebug", thinking
  it would be easier to understand the priority scheme this way, but abandoned
  that approach midway.  Which is easier to grasp?

Tested on x86_64-linux.