[gdb] Fix reverse stepping to first line in optimized function
Consider the test-case contained in this patch, compiled with -O2 on
x86_64-linux with gcc 7.5.0.

When forward-stepping into foo, we land at the first line in foo:
...
32        foo ();
(gdb) step
foo () at step-reverse-opt.c:24
24        i = x;
...

Now let's try to forward-step away from the first line:
...
(gdb) step
25        j = y;
...
That went fine.

Now let's try to reverse-step back to the first line:
...
(gdb) reverse-step

No more reverse-execution history.
main () at step-reverse-opt.c:32
32        foo ();
(gdb)
...
That didn't work.

The problem is as follows.

The insns of foo are:
...
00000000004004c0 <foo>:
  4004c0: 8b 05 66 1b 00 00       mov    0x1b66(%rip),%eax        # 40202c <x>
  4004c6: 89 05 64 1b 00 00       mov    %eax,0x1b64(%rip)        # 402030 <i>
  4004cc: 8b 05 66 1b 00 00       mov    0x1b66(%rip),%eax        # 402038 <y>
  4004d2: 89 05 4c 1b 00 00       mov    %eax,0x1b4c(%rip)        # 402024 <j>
  4004d8: 8b 05 4a 1b 00 00       mov    0x1b4a(%rip),%eax        # 402028 <z>
  4004de: 89 05 50 1b 00 00       mov    %eax,0x1b50(%rip)        # 402034 <k>
  4004e4: c3                      ret
...
and the line table in GDB is:
...
INDEX  LINE   ADDRESS            IS-STMT PROLOGUE-END
4      23     0x00000000004004c0 Y
5      24     0x00000000004004c0 Y
6      25     0x00000000004004cc Y
7      26     0x00000000004004d8 Y
8      27     0x00000000004004e4 Y
9      END    0x00000000004004e5 Y
...

When stopped at line 25, we'd expect the reverse-step to bring us back to the
insn at 0x4004c0 with line number 24.  However, we don't stop because
this piece of code in process_event_stop_test triggers:
...
      /* When stepping backward, stop at beginning of line range
	 (unless it's the function entry point, in which case
	 keep going back to the call point).  */
      CORE_ADDR stop_pc = ecs->event_thread->stop_pc ();
      if (stop_pc == ecs->event_thread->control.step_range_start
	  && stop_pc != ecs->stop_func_start
	  && execution_direction == EXEC_REVERSE)
	end_stepping_range (ecs);
      else
	keep_going (ecs);
...

The comment describes exactly what happens, which is the correct thing to
do when stepping through a prologue, but in this case the function has no
prologue, and so "keep going back to the call point" is the wrong thing to do.

Fix this by using "skip_prologue_using_sal (gdbarch, stop_pc) == stop_pc" to
check whether the function has a prologue or not.

Tested on x86_64-linux.

PR gdb/29996
Bug: https://sourceware.org/bugzilla/show_bug.cgi?id=29996