[gdb/testsuite] Simplify gdb.base/unwind-on-each-insn.exp
In test-case gdb.base/unwind-on-each-insn.exp, we try to determine the last
insn in function foo.

This in it self is fragile, as demonstrated by commit 91836f41e20 ("Powerpc
fix for gdb.base/unwind-on-each-insn.exp").

But the purpose of finding the last insn is to stop stepping in foo when
arriving at that last insn.

There is however no guarantee that:
- the last insn is actually executed, nor
- that the last insn is executed last, nor
- that the last insn is executed once.

Fix this by simplying the test-case to continue stepping till stepping out of
foo.

Tested on x86_64-linux.