gdb-testsuite-add-kfail-in-gdb.threads-fork-plus-threads.exp
[gdb/testsuite] Add KFAIL in gdb.threads/fork-plus-threads.exp

When running test-case gdb.threads/fork-and-threads.exp on a VM with openSUSE
Tumbleweed, with the VM bound to 1 cpu with 75% execution cap, I get:
...
(gdb) info inferiors^M
  Num  Description       Connection           Executable        ^M
* 1    <null>                                 fork-plus-threads ^M
  11   <null>                                 fork-plus-threads ^M
(gdb) FAIL: gdb.threads/fork-plus-threads.exp: detach-on-fork=off: \
  only inferior 1 left
...

The test checks that all removable inferiors are indeed removed from the
inferior list after exit, and evidently this didn't happen for inferior 11
(which is added by fork rather than a user command, and therefore removable).

I've investigated why that is that case, and it's because its refcount didn't
drop to 0.

This seems like a bug to me, so add a KFAIL for this.

Tested on x86_64-linux.