gdb-testsuite-work-around-skip_prologue-problems-in-gdb.threads-process-dies-while-detaching.exp
[gdb/testsuite] Work around skip_prologue problems in gdb.threads/process-dies-while-detaching.exp

On powerpc64le-linux, I run into:
...
[Inferior 1 (process 5156) exited normally]^M
(gdb) FAIL: gdb.threads/process-dies-while-detaching.exp: single-process: \
  detach: detach: continue to breakpoint: _exit (the program exited)
...

What happens is the following:
- a breakpoint is set on _exit,
- a continue is issued
- the continue is supposed to hit the breakpoint, but instead
  the program exits.

I traced this down to the breakpoint on _exit being set too far from function
entry.  This is caused by the skip_prologue function (in rs6000-tdep.c)
optimistically ignoring insns it doesn't recognize.  In particular, it walks
past the system call instruction "sc" which initiates the actual exit.

While this needs fixing, we don't want to be testing this behaviour in this
test-case.

[ Initially I tried to fix it by setting a breakpoint on "*_exit" instead, but
that one only sets one location.  The breakpoint on "_exit" sets two
locations, one in /lib64/libc.so.6 and one in /lib64/ld64.so.2.  I tried on
x86_64 and there the breakpoint on "*_exit" mapped to the /lib64/libc.so.6
location, and the test-case passed.  But on powerpc it mapped to the
/lib64/ld64.so.2 location and I still got the same failures. ]

Fix this by setting two breakpoints on the calls to _exit and exit instead.

Tested on x86_64-linux and powerpc64le-linux.