[gdb/symtab] Resolve deferred entries, inter-shard case
When running test-case gdb.cp/breakpoint-locs.exp with target board
cc-with-dwz, we occasionally run into:
...
FAIL: gdb.cp/breakpoint-locs.exp: break N1::C1::baz
...
due to the breakpoint having one instead of two locations.

By using "maint set worker-threads 0" we can make the test-case fail reliably.

In the failing case the inlined_subroutine DIE at 0x1cf has an incorrect
parent, and the one at 0x148 has the correct parent:
...
    [7] ((cooked_index_entry *) 0x7fdf800030b0)
    name:       baz
    canonical:  baz
    qualified:  N1::C1::baz
    DWARF tag:  DW_TAG_inlined_subroutine
    flags:      0x0 []
    DIE offset: 0x148
    parent:     ((cooked_index_entry *) 0x7fdf80002ed0) [C1]
  ...
    [20] ((cooked_index_entry *) 0x7fdf84005340)
    name:       baz
    canonical:  baz
    qualified:  baz
    DWARF tag:  DW_TAG_inlined_subroutine
    flags:      0x0 []
    DIE offset: 0x1cf
    parent:     ((cooked_index_entry *) 0)
...
and in the passing case it's the other way around.

The corresponding dwarf for 0x148 is this:
...
 <0><b>: Abbrev Number: 34 (DW_TAG_partial_unit)
 <1><14>: Abbrev Number: 37 (DW_TAG_namespace)
    <15>   DW_AT_name        : N1
 <2><19>: Abbrev Number: 36 (DW_TAG_class_type)
    <1a>   DW_AT_name        : C1
 <3><20>: Abbrev Number: 32 (DW_TAG_subprogram)
    <21>   DW_AT_external    : 1
    <21>   DW_AT_name        : baz
    <27>   DW_AT_linkage_name: _ZN2N12C13bazEv
    <2b>   DW_AT_accessibility: 1       (public)
    <2c>   DW_AT_declaration : 1
 <1><2e>: Abbrev Number: 35 (DW_TAG_subprogram)
    <2f>   DW_AT_specification: <0x20>
    <30>   DW_AT_inline      : 3        (declared as inline and inlined)
 <0><ee>: Abbrev Number: 16 (DW_TAG_compile_unit)
    <f3>   DW_AT_language    : 4        (C++)
    <f4>   DW_AT_name        : gdb.cp/breakpoint-locs.cc
 <1><13b>: Abbrev Number: 25 (DW_TAG_subprogram)
    <13c>   DW_AT_specification: <0x115>
    <13d>   DW_AT_low_pc      : 0x4004d7
    <145>   DW_AT_high_pc     : 16
    <146>   DW_AT_frame_base  : 1 byte block: 9c        (DW_OP_call_frame_cfa)
    <148>   DW_AT_GNU_all_call_sites: 1
 <2><148>: Abbrev Number: 24 (DW_TAG_inlined_subroutine)
    <149>   DW_AT_abstract_origin: <0x2e>
    <14d>   DW_AT_low_pc      : 0x4004db
    <155>   DW_AT_high_pc     : 9
    <156>   DW_AT_call_file   : 1
    <157>   DW_AT_call_line   : 22
...
and for 0x1cf it's similar but in a CU with name gdb.cp/breakpoint-locs-2.cc.

In the target board unix case, the test-case passes because both CUs are
expanded, because each CU contains the equivalent of 0x2e.

But in the dwz case, dwz factored out that DIE into the PU, so that mechanism
no longer works, and instead we rely on the DW_TAG_inlined_subroutine DIEs to
trigger the expansion, which requires that the parent of each DIE is correct.

[ Note that also in the partial symtab we relied on the
DW_TAG_inlined_subroutine DIEs, see commit f9b5d5ea18a ("[gdb/symtab] Fix
missing breakpoint location for inlined function"). ]

By switching on debug_handle_deferred_entries, we get:
...
  0x0000000000000015 0x2708cf0 (0x14)
  0x000000000000001a 0x2708d20 (0x19)
  0x000000000000002d 0x2708cf0 (0x14)
  0x000000000000002e 0x2708d20 (0x19)
  0x000000000000002f 0x2708d50 (0x2e)
  0x0000000000000039 0x0
  0x000000000000010f 0x2708e10 (0x10e)
  0x0000000000000121 0x0
  0x000000000000013b 0x2708e10 (0x10e)
  0x000000000000013c 0x2708ea0 (0x13b)
  0x0000000000000148 0x2708d20 (0x19)
  0x0000000000000149 0x2708ea0 (0x13b)
  0x000000000000016d 0x0
  0x000000000000019a 0x2708f60 (0x199)
  0x00000000000001ac 0x0
  0x00000000000001c2 0x2708f60 (0x199)
  0x00000000000001c3 0x2708ff0 (0x1c2)
  0x00000000000001f4 0x0
...

So, 0x148 has parent 0x19, in other words N1::C1, and 0x1cf, in other words bar.
The parent of 0x1cf is incorrect, but it was never used.  Instead, at the point
the DIE was indexed, find_parent was called, which returned nullptr, and that
was used instead.  Which is also incorrect.

The problem is that m_die_range_map only has the scope of a parsing a single CU.
For the first CU, that includes the PU as well, but not for the second.

[ Then there's the problem that find_parent cannot distinguish between the
"no parent" and "don't know" cases, which means that getting the incorrect
parent happens silently. ]

We need to generically solve the case of inter-CU dependencies, and that includes
inter-shard dependencies as well.

So we:
- mark deferred entries in m_die_range_map using (void *)-1, to prevent
  incorrect entries for deferred dies,
- defer all inter-CU dependencies (and consequently both 0x148 and 0x1cf are
  deferred),
- move m_die_range_map and m_deferred_dies to cooked_index_shard, and
- move handle_deferred_dies to the cooked_index, where it is called in the
  constructor, and update it to handle the intra-shard case,
such that we get:
...
  0x0000000000000015 0x4623d10 (0x14)
  0x000000000000001a 0x4623d40 (0x19)
  0x000000000000002d 0x4623d10 (0x14)
  0x000000000000002e 0x4623d40 (0x19)
  0x000000000000002f 0x4623d70 (0x2e)
  0x0000000000000039 0x0
  0x000000000000010f 0x4623e30 (0x10e)
  0x0000000000000121 0x0
  0x000000000000013b 0x4623e30 (0x10e)
  0x000000000000013c 0x4623ec0 (0x13b)
  0x0000000000000148 0xffffffffffffffff (deferred)
  0x0000000000000149 0x4623ec0 (0x13b)
  0x000000000000016d 0x0
  0x000000000000019a 0x4623f50 (0x199)
  0x00000000000001ac 0x0
  0x00000000000001c2 0x4623f50 (0x199)
  0x00000000000001c3 0x4623fe0 (0x1c2)
  0x00000000000001cf 0xffffffffffffffff (deferred)
  0x00000000000001d0 0x4623fe0 (0x1c2)
  0x00000000000001f4 0x0
Resolve deferred: 0x148 -> 0x2e: 0x19
Resolve deferred: 0x1cf -> 0x2e: 0x19
...

Tested on x86_64-linux.